Move packages to new workspace for cleaner management
