+++ package.json
@@ -12,21 +12,12 @@
     "dev:vite": "npm -C ./packages/example-vite run dev",
     "build": "npm -C ./packages/reactivue run build && cp README.md ./packages/reactivue",
     "release": "cd ./packages/reactivue && npm run build && npx bumpp --commit --tag --push && npm publish",
-    "test": "jest"
+    "test": "npm -C ./packages/tests run test"
   },
   "devDependencies": {
     "@antfu/eslint-config-ts": "^0.4.3",
-    "@testing-library/jest-dom": "^5.11.8",
-    "@testing-library/react": "^11.2.3",
-    "@types/jest": "^26.0.20",
-    "babel-jest": "^26.6.3",
     "eslint": "^7.12.1",
-    "eslint-plugin-react": "^7.21.5",
-    "jest": "^26.6.3",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "ts-jest": "^26.4.4",
-    "ts-node": "^9.1.1"
+    "eslint-plugin-react": "^7.21.5"
   },
   "eslintConfig": {
     "extends": [
