+++ packages/tests/src/computed.spec.tsx
@@ -1,7 +1,8 @@
+// eslint-disable-next-line no-use-before-define
 import React from 'react'
 import { render } from '@testing-library/react'
 import { screen, waitFor } from '@testing-library/dom'
-import { useSetup, toRef, computed } from '../src'
+import { useSetup, toRef, computed } from 'reactivue'
 
 const ComputedTest = (Props: { hello: string }) => {
   const { comp } = useSetup((props) => {
