+++ packages/tests/src/defineComponent.spec.tsx
@@ -1,23 +1,21 @@
+// eslint-disable-next-line no-use-before-define
 import React from 'react'
 import { render } from '@testing-library/react'
 import { screen, waitFor } from '@testing-library/dom'
-import { createSetup, ref } from '../src'
+import { defineComponent, ref } from 'reactivue'
 
-const useMsg = createSetup(() => {
+const DefineTest = defineComponent(() => {
   const msg = ref('Hello, world!')
 
   return {
     msg,
   }
+}, ({ msg }) => {
+  return <p>{msg}</p>
 })
 
-const CreateSetupTest = () => {
-  const { msg } = useMsg()
-  return <p>{msg}</p>
-}
-
-it('should render basic createSetup return', async() => {
-  render(<CreateSetupTest/>)
+it('should render basic defineComponent component', async() => {
+  render(<DefineTest/>)
 
   await waitFor(() => {
     const el = screen.getByText('Hello, world!')
