+++ packages/tests/src/useSetup.spec.tsx
@@ -1,7 +1,8 @@
+// eslint-disable-next-line no-use-before-define
 import React from 'react'
 import { render } from '@testing-library/react'
 import { screen, waitFor } from '@testing-library/dom'
-import { useSetup, ref, toRef } from '../src'
+import { useSetup, ref, toRef } from 'reactivue'
 
 const SetupTest = (Props: { hello?: string }) => {
   const { msg, other } = useSetup((props) => {
