+++ packages/tests/src/watch.spec.tsx
@@ -1,7 +1,8 @@
+// eslint-disable-next-line no-use-before-define
 import React from 'react'
 import { render } from '@testing-library/react'
 import { waitFor } from '@testing-library/dom'
-import { useSetup, toRef, watch, watchEffect } from '../src'
+import { useSetup, toRef, watch, watchEffect } from 'reactivue'
 
 const watchJestFn = jest.fn()
 const watchEffectJestFn = jest.fn()
